flattened modules
