+++ external-crates/move/move-execution/v2/crates/move-stdlib-natives/src/helpers.rs
