+++ external-crates/move/move-execution/v2/crates/move-stdlib-natives/src/string.rs
